refactor(validation):优化投币行为校验逻辑并移除冗余代码

+++ only-danmuku-application/src/main/kotlin/edu/only4/danmuku/application/subscribers/domain/customer_action/CustomerCoinGivenDomainEventSubscriber.kt
@@ -19,8 +19,8 @@ class CustomerCoinGivenDomainEventSubscriber {
         val action = event.entity
         Mediator.commands.send(
             TransferCoinCmd.Request(
-                senderUserId = action.customerId.toLong(),
-                receiverUserId = action.videoOwnerId.toLong(),
+                senderUserId = action.customerId,
+                receiverUserId = action.videoOwnerId,
                 amount = action.actionCount
             )
         )
@@ -33,7 +33,7 @@ class CustomerCoinGivenDomainEventSubscriber {
 
         Mediator.commands.send(
             UpdateVideoStatisticsCmd.Request(
-                videoId = action.videoId.toLong(),
+                videoId = action.videoId,
                 coinCountDelta = action.actionCount
             )
         )
